+++ src/main/java/com/innovation/stockstock/account/domain/Account.java
@@ -26,7 +26,7 @@ public class Account extends Timestamped {
     @Column(name = "account_id")
     private Long id;
 
-    private int accountNumber;
+    private Long accountNumber;
 
     private int seedMoney;
